+++ src/app/features/services/manage-user-api.service.ts
@@ -6,7 +6,7 @@ import { Observable } from 'rxjs';
 @Injectable({
   providedIn: 'root'
 })
-export class ApiService {
+export class ManageUserApiService {
   basePath = 'user';
   constructor(private apiConfigService: ApiConfigService) {}